feat: sync hh statuses
